- Corrected missing root link for sitemap (02/04/2025) - Corrected Update of Product (02/04/2025) - Removed use of StripeVersion (02/04/2025) - Moved ApiKey to bundle.yaml (02/04/2025) - Added Webhook support (02/04/2025) - Removed unsued Payment fields (02/04/2025) - Added Payment->stripeMethod (02/04/2025) - Added link to Stripe payment from dashboard (02/04/2025)

+++ src/Listener/Traits/MediaTrait.php
@@ -25,10 +25,10 @@ trait MediaTrait
         }
     }
 
-    // Delete media
+    // Delete media (if media is not deleted, use MediaDeleteCommand)
     public function deleteMedia($entity): void
     {
-        if (method_exists($entity, 'getFile') && null !== $entity->getFile()) {
+        if (method_exists($entity, 'getName') && method_exists($entity, 'getFile') && null !== $entity->getFile()) {
             $name = __DIR__ . self::MEDIA_ROOT . $entity->getName();
             if (file_exists($name)) {
                 unlink($name);
